worker/requirements.txt
flask==3.0.0
google-cloud-storage==2.14.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==3.0.1